[refact][feat] 20240709_Patch_Note * Coredata   - 객체별 Set 과정중 nil 데이터 확인로직 추가   - 배열 데이터를 get or delete 객체에 대해서 빈 배열일 경우, 조기종료 로직추가 * 회원가입   - local storage(Coredata) set 로직강화 * 로그인   - FirebaseAuth 에는 등록되어 있으나, Server 에 저장정보가 없는 사용자 처리로직 추가     : 회원가입 중 앱종료, 혹은 회원가입 취소 사용자 대상     : 진행된 과정 초기화 및 FirebaseAuth 등록유저 제거 * LoginService   - 분산되어있는 로그아웃 및 회원탈퇴 코드를 이곳에 배치 * Etc   - NetworkManager 동작방식 변경   - 일부 View 내 'showAlert' 변수명 변경   - 일부 부정확한 Error 관련 print 로그명 변경

+++ teamplan/Sources/Repository/Log/Access/AccessLogServicesCoredata.swift
@@ -20,7 +20,7 @@ final class AccessLogServicesCoredata: AccessLogObjectManage {
         self.context = LocalStorageManager.shared.context
     }
     
-    func setObject(with object: Object) {
+    func setObject(with object: Object) -> Bool {
         createEntity(with: object)
     }
     
@@ -33,11 +33,23 @@ final class AccessLogServicesCoredata: AccessLogObjectManage {
     // Get: Full Log
     func getFullObjects(with userId: String) throws -> [Object] {
         let entities = try getFullEntity(with: userId)
+        
+        if entities.isEmpty {
+            print("[AccessLogRepo] There is no log to get")
+            return []
+        }
+        
         return try convertToObjects(with: entities)
     }
     
     func deleteObject(with userId: String) throws {
         let entities = try getFullEntity(with: userId)
+        
+        if entities.isEmpty {
+            print("[AccessLogRepo] There is no log to delete")
+            return
+        }
+        
         for entity in entities {
             self.context.delete(entity)
         }
@@ -63,11 +75,22 @@ extension AccessLogServicesCoredata {
 
 extension AccessLogServicesCoredata {
     
-    private func createEntity(with log: Object) {
+    private func createEntity(with log: Object) -> Bool {
         let entity = Entity(context: context)
         
         entity.user_id = log.userId
         entity.access_record = log.accessRecord
+        
+        // Optional property nil checks
+        if entity.user_id == nil {
+            print("[AccessLogRepo] nil detected: 'user_id'")
+            return false
+        }
+        if entity.access_record == nil {
+            print("[AccessLogRepo] nil detected: 'access_record'")
+            return false
+        }
+        return true
     }
     
     private func getFullFetchRequest(with userId: String, sortNeed: Bool) -> NSFetchRequest<Entity> {
